Refactored authProvider, deleted unused files
