+++ pharmd-app/src/services/Auth/authError.js
@@ -1,0 +1,7 @@
+export default ({ status }) => {
+    if (status === 401 || status === 403) {
+        localStorage.removeItem("token");
+        return Promise.reject();
+    }
+    return Promise.resolve();
+}
