+++ pharmd-app/src/services/Auth/getPermissions.js
@@ -1,0 +1,4 @@
+export default () => {
+    const role = localStorage.getItem("permissions");
+    return role ? Promise.resolve(role) : Promise.reject();
+}
