+++ pharmd-app/src/services/Auth/isAuth.js
@@ -1,0 +1,3 @@
+export default () => {
+    return localStorage.getItem("token") ? Promise.resolve() : Promise.reject();
+}
