+++ pharmd-app/src/services/Auth/logout.js
@@ -1,0 +1,6 @@
+export default () => {
+    localStorage.removeItem("token");
+    localStorage.removeItem("permissions");
+    localStorage.removeItem("userInfo");
+    return Promise.resolve();
+}
