+++ pharmd-app/src/services/index.js
@@ -1,12 +1,8 @@
-import CourseService from "./CourseService";
 import FileUploadService from "./FileUploadService";
-import StudentsService from "./StudentsService";
-import GeneralService from "./GeneralService";
+import AuthProvider from './Auth';
 
 // This should be kept up-to-date as new service files are added
-export default {
-  CourseService,
+export {
   FileUploadService,
-  StudentsService,
-  GeneralService
+  AuthProvider
 };
