--- pharmd-app/src/styles/globalStyles.css
@@ -1,19 +1,0 @@
-/*Note: Many of these styles are temporary until Jose's design system is fully integrated*/
-
-.pharmd-button {
-  padding: 0.5rem;
-  border-radius: 15%;
-  background-color: #61dafb;
-  box-shadow: #282c34;
-  color: #282c34;
-  font-weight: bold;
-  border-style: none;
-}
-
-.pharmd-button:active {
-  background-color: #3f88a1;
-}
-
-.text-center {
-  text-align: center;
-}
